Slave upd

+++ editor/Editor.c
@@ -1,7 +1,7 @@
 #include "Editor.h"
 
 
-struct Editor set_Editor(int id, char* name, float productivity){
+struct Editor set_Editor(int id, char* name, double productivity){
     struct Editor e;
     e.id;
     e.name = name;
