+++ editor/Editor.h
@@ -2,9 +2,10 @@ struct Editor
 {
     int id;
     char *name;
-    float productivity;
+    double productivity;
 };
 
 
-struct Editor set_Editor(int,char*,float);
+struct Editor set_Editor(int,char*,double);
+
